Split input handling
