+++ input.c
@@ -1,0 +1,348 @@
+#define _XOPEN_SOURCE_EXTENDED 1
+#define _XOPEN_SOURCE 600
+#include <stdlib.h>
+#include <time.h>
+#include <wchar.h>
+#include <ncursesw/curses.h>
+
+#include "undo.h"
+#include "input.h"
+#include "screen.h"
+#include "text.h"
+
+#define max(a, b) ((a) > (b) ? (a) : (b))
+/* globals from ee.c */
+extern int in;
+extern int scr_vert;
+extern int scr_horz;
+extern int horiz_offset;
+extern int last_line;
+extern int gold;
+extern char *ascii_code_str;
+extern WINDOW *text_win;
+extern ee_char *point;
+
+/* helpers from ee.c */
+char *get_string(const char *prompt, int advance);
+void bottom(void);
+void command_prompt(void);
+void down(void);
+void search_prompt(void);
+void redo_action(void);
+void bol(void);
+void delete(int disp);
+void insert_line(int disp);
+void undo_action(void);
+void left(int disp);
+void insert(int character);
+void move_rel(int direction, int lines);
+void eol(void);
+void top(void);
+void up(void);
+void undel_word(void);
+void del_word(void);
+int search(int display_message);
+void del_line(void);
+void undel_line(void);
+struct menu_entries {
+    char *item_string;
+    int (*procedure)(struct menu_entries *);
+    struct menu_entries *ptr_argument;
+    int (*iprocedure)(int);
+    void (*nprocedure)(void);
+    int argument;
+};
+int menu_op(struct menu_entries *);
+extern struct menu_entries main_menu[];
+void adv_word(void);
+void adv_line(void);
+void paint_info_win(void);
+void midscreen(int line, ee_char *pnt);
+int collect_input_chunk(int *buf, int max)
+{
+    int len = 0;
+    wint_t ch;
+    int rc = wget_wch(text_win, &ch);
+    if (rc == ERR)
+        return 0;
+    buf[len++] = (int)ch;
+
+    nodelay(text_win, TRUE);
+    struct timespec delay = {0, 30000000};
+    while (len < max) {
+        rc = wget_wch(text_win, &ch);
+        if (rc == ERR) {
+            nanosleep(&delay, NULL);
+            rc = wget_wch(text_win, &ch);
+            if (rc == ERR)
+                break;
+        }
+        buf[len++] = (int)ch;
+    }
+    nodelay(text_win, FALSE);
+    return len;
+}
+
+void start_action(void)
+{
+    undo_begin_chunk();
+}
+
+
+void
+control(void)
+{
+	char *string;
+
+	if (in == 1)		/* control a	*/
+	{
+		string = get_string(ascii_code_str, TRUE);
+		if (*string != '\0')
+		{
+			in = atoi(string);
+			wmove(text_win, scr_vert, (scr_horz - horiz_offset));
+			insert(in);
+		}
+		free(string);
+	}
+	else if (in == 2)	/* control b	*/
+		bottom();
+	else if (in == 3)	/* control c	*/
+	{
+		command_prompt();
+	}
+	else if (in == 4)	/* control d	*/
+		down();
+	else if (in == 5)	/* control e	*/
+		search_prompt();
+	else if (in == 6)	/* control f	*/
+		redo_action();
+	else if (in == 7)	/* control g	*/
+		bol();
+	else if (in == 8)	/* control h	*/
+		delete(TRUE);
+	else if (in == 9)	/* control i	*/
+		;
+	else if (in == 10)	/* control j	*/
+		insert_line(TRUE);
+	else if (in == 11)	/* control k	*/
+		undo_action();
+	else if (in == 12)	/* control l	*/
+		left(TRUE);
+	else if (in == 13)	/* control m	*/
+		insert_line(TRUE);
+	else if (in == 14)	/* control n	*/
+		move_rel('d', max(5, (last_line - 5)));
+	else if (in == 15)	/* control o	*/
+		eol();
+	else if (in == 16)	/* control p	*/
+		move_rel('u', max(5, (last_line - 5)));
+	else if (in == 17)	/* control q	*/
+		;
+	else if (in == 18)	/* control r	*/
+		right(TRUE);
+	else if (in == 19)	/* control s	*/
+		;
+	else if (in == 20)	/* control t	*/
+		top();
+	else if (in == 21)	/* control u	*/
+		up();
+	else if (in == 22)	/* control v	*/
+		undel_word();
+	else if (in == 23)	/* control w	*/
+		del_word();
+	else if (in == 24)	/* control x	*/
+		search(TRUE);
+	else if (in == 25)	/* control y	*/
+		del_line();
+	else if (in == 26)	/* control z	*/
+		undel_line();
+	else if (in == 27)	/* control [ (escape)	*/
+	{
+		menu_op(main_menu);
+	}	
+}
+/*
+ |      Emacs control-key bindings
+ */
+
+void
+emacs_control(void)
+{
+	char *string;
+
+	if (in == 1)		/* control a	*/
+		bol();
+	else if (in == 2)	/* control b	*/
+		left(TRUE);
+	else if (in == 3)	/* control c	*/
+	{
+		command_prompt();
+	}
+	else if (in == 4)	/* control d	*/
+		undo_action();
+	else if (in == 5)	/* control e	*/
+		eol();
+	else if (in == 6)	/* control f	*/
+		right(TRUE);
+	else if (in == 7)	/* control g	*/
+		move_rel('u', max(5, (last_line - 5)));
+	else if (in == 8)	/* control h	*/
+		delete(TRUE);
+	else if (in == 9)	/* control i	*/
+		;
+	else if (in == 10)	/* control j	*/
+		redo_action();
+	else if (in == 11)	/* control k	*/
+		del_line();
+	else if (in == 12)	/* control l	*/
+		undel_line();
+	else if (in == 13)	/* control m	*/
+		insert_line(TRUE);
+	else if (in == 14)	/* control n	*/
+		down();
+	else if (in == 15)	/* control o	*/
+	{
+		string = get_string(ascii_code_str, TRUE);
+		if (*string != '\0')
+		{
+			in = atoi(string);
+			wmove(text_win, scr_vert, (scr_horz - horiz_offset));
+			insert(in);
+		}
+		free(string);
+	}
+	else if (in == 16)	/* control p	*/
+		up();
+	else if (in == 17)	/* control q	*/
+		;
+	else if (in == 18)	/* control r	*/
+		undel_word();
+	else if (in == 19)	/* control s	*/
+		;
+	else if (in == 20)	/* control t	*/
+		top();
+	else if (in == 21)	/* control u	*/
+		bottom();
+	else if (in == 22)	/* control v	*/
+		move_rel('d', max(5, (last_line - 5)));
+	else if (in == 23)	/* control w	*/
+		del_word();
+	else if (in == 24)	/* control x	*/
+		search(TRUE);
+	else if (in == 25)	/* control y	*/
+		search_prompt();
+	else if (in == 26)	/* control z	*/
+		adv_word();
+	else if (in == 27)	/* control [ (escape)	*/
+	{
+		menu_op(main_menu);
+	}	
+}
+
+
+void
+function_key(void)
+{
+	if (in == KEY_LEFT)
+		left(TRUE);
+	else if (in == KEY_RIGHT)
+		right(TRUE);
+	else if (in == KEY_HOME)
+		bol();
+	else if (in == KEY_END)
+		eol();
+	else if (in == KEY_UP)
+		up();
+	else if (in == KEY_DOWN)
+		down();
+	else if (in == KEY_NPAGE)
+		move_rel('d', max( 5, (last_line - 5)));
+	else if (in == KEY_PPAGE)
+		move_rel('u', max(5, (last_line - 5)));
+	else if (in == KEY_DL)
+		del_line();
+	else if (in == KEY_DC)
+		undo_action();
+	else if (in == KEY_BACKSPACE)
+		delete(TRUE);
+	else if (in == KEY_IL)
+	{		/* insert a line before current line	*/
+		insert_line(TRUE);
+		left(TRUE);
+	}
+	else if (in == KEY_F(1))
+		gold = !gold;
+	else if (in == KEY_F(2))
+	{
+		if (gold)
+		{
+			gold = FALSE;
+			undel_line();
+		}
+		else
+			redo_action();
+	}
+	else if (in == KEY_F(3))
+	{
+		if (gold)
+		{
+			gold = FALSE;
+			undel_word();
+		}
+		else
+			del_word();
+	}
+	else if (in == KEY_F(4))
+	{
+		if (gold)
+		{
+			gold = FALSE;
+			paint_info_win();
+			midscreen(scr_vert, point);
+		}
+		else
+			adv_word();
+	}
+	else if (in == KEY_F(5))
+	{
+		if (gold)
+		{
+			gold = FALSE;
+			search_prompt();
+		}
+		else
+			search(TRUE);
+	}
+	else if (in == KEY_F(6))
+	{
+		if (gold)
+		{
+			gold = FALSE;
+			bottom();
+		}
+		else
+			top();
+	}
+	else if (in == KEY_F(7))
+	{
+		if (gold)
+		{
+			gold = FALSE;
+			eol();
+		}
+		else
+			bol();
+	}
+	else if (in == KEY_F(8))
+	{
+		if (gold)
+		{
+			gold = FALSE;
+			command_prompt();
+		} 
+		else
+                adv_line();
+        }
+}
+
